resolverを分割
