+++ mock/graphql/lib/graphql/types/userType.graphql
